[#128] Removed templating plugin, and referenced new plugin on Maven Central
