+++ clients/java-client-lib/client-command-builders/pom.xml
@@ -26,12 +26,11 @@
             <plugin>
                 <groupId>net.zscript.maven-templates</groupId>
                 <artifactId>mustache-templating-maven-plugin</artifactId>
-                <version>0.0.1-SNAPSHOT</version>
 
                 <dependencies>
                     <dependency>
                         <groupId>net.zscript</groupId>
-                        <artifactId>zscript-model-mustache-templating-adapter</artifactId>
+                        <artifactId>zscript-model-mustache-templating-context-loader</artifactId>
                         <version>${project.version}</version>
                     </dependency>
                     <dependency>
@@ -45,7 +44,7 @@
                     <templateDirectory>${project.resources[0].directory}/templates/commands</templateDirectory>
                     <mainTemplate>JavaCommandBuilder.mustache</mainTemplate>
                     <outputDirectory>${project.build.directory}/generated-sources/zscript</outputDirectory>
-                    <templatingMapperClass>net.zscript.model.templating.adapter.ZscriptModuleTemplatingPluginMapper</templatingMapperClass>
+                    <contextLoaderClass>net.zscript.model.templating.contextloader.ZscriptModuleTemplatingContextLoader</contextLoaderClass>
                 </configuration>
 
                 <executions>
