+++ model/model-templating-context-loader/pom.xml
@@ -1,0 +1,42 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>net.zscript</groupId>
+        <artifactId>zscript-model</artifactId>
+        <version>0.0.1-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>zscript-model-mustache-templating-context-loader</artifactId>
+    <name>Zscript Model Generator Mustache Context Loader</name>
+
+    <description>
+        This is a "build" project that builds a component for reading the Zscript model into a Mustache template.
+
+        Note, it can have dependencies that the production code shouldn't, like Guava.
+    </description>
+
+    <properties>
+        <version.guava>32.1.2-jre</version.guava>
+    </properties>
+
+    <dependencies>
+        <dependency>
+            <groupId>net.zscript.maven-templates</groupId>
+            <artifactId>mustache-templating-context-loader</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>net.zscript</groupId>
+            <artifactId>zscript-model-schema</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.google.guava</groupId>
+            <artifactId>guava</artifactId>
+            <version>${version.guava}</version>
+        </dependency>
+    </dependencies>
+
+</project>
