+++ pom.xml
@@ -33,18 +33,18 @@
         <version.jimfs>1.3.0</version.jimfs>
         <version.equalsverifier>3.15.1</version.equalsverifier>
         <version.apachecommons>2.13.0</version.apachecommons>
-        <version.snakeyaml>2.2</version.snakeyaml>
 
         <!-- plugin dependency versions -->
         <!--        <version.nar-maven-plugin>3.10.2-SNAPSHOT</version.nar-maven-plugin>-->
         <version.nar-maven-plugin>3.10.2-alpha</version.nar-maven-plugin>
-        <version.jacoco-maven-plugin>0.8.7</version.jacoco-maven-plugin>
+        <version.jacoco-maven-plugin>0.8.10</version.jacoco-maven-plugin>
         <version.exec-maven-plugin>3.0.0</version.exec-maven-plugin>
         <version.maven-compiler-plugin>3.10.1</version.maven-compiler-plugin>
         <version.maven-surefire-plugin>3.0.0-M5</version.maven-surefire-plugin>
         <version.maven-failsafe-plugin>3.0.0-M5</version.maven-failsafe-plugin>
         <version.maven-assembly-plugin>3.4.2</version.maven-assembly-plugin>
         <version.maven-enforcer-plugin>3.4.2</version.maven-enforcer-plugin>
+        <version.mustache-templating-maven-plugin>0.0.1</version.mustache-templating-maven-plugin>
     </properties>
 
     <modules>
@@ -89,9 +89,9 @@
                 <version>${version.antlr}</version>
             </dependency>
             <dependency>
-                <groupId>org.yaml</groupId>
-                <artifactId>snakeyaml</artifactId>
-                <version>${version.snakeyaml}</version>
+                <groupId>net.zscript.maven-templates</groupId>
+                <artifactId>mustache-templating-context-loader</artifactId>
+                <version>${version.mustache-templating-maven-plugin}</version>
             </dependency>
 
 
@@ -155,6 +155,11 @@
                     <extensions>true</extensions>
                 </plugin>
                 <plugin>
+                    <groupId>net.zscript.maven-templating</groupId>
+                    <artifactId>mustache-templating-maven-plugin</artifactId>
+                    <version>${version.mustache-templating-maven-plugin}</version>
+                </plugin>
+                <plugin>
                     <groupId>org.jacoco</groupId>
                     <artifactId>jacoco-maven-plugin</artifactId>
                     <version>${version.jacoco-maven-plugin}</version>
@@ -203,7 +208,7 @@
 
     <organization>
         <name>Zscript Team</name>
-        <url>https://groups.google.com/g/zscript-group</url>
+        <url>http://zscript.net/</url>
     </organization>
 
     <licenses>
