+++ receivers/native/cpp-model-components/pom.xml
@@ -23,12 +23,11 @@
             <plugin>
                 <groupId>net.zscript.maven-templates</groupId>
                 <artifactId>mustache-templating-maven-plugin</artifactId>
-                <version>0.0.1-SNAPSHOT</version>
 
                 <dependencies>
                     <dependency>
                         <groupId>net.zscript</groupId>
-                        <artifactId>zscript-model-mustache-templating-adapter</artifactId>
+                        <artifactId>zscript-model-mustache-templating-context-loader</artifactId>
                         <version>${project.version}</version>
                     </dependency>
                     <dependency>
@@ -45,7 +44,7 @@
 
                 <configuration>
                     <outputDirectory>${generated-code-path}</outputDirectory>
-                    <templatingMapperClass>net.zscript.model.templating.adapter.ZscriptModuleTemplatingPluginMapper</templatingMapperClass>
+                    <contextLoaderClass>net.zscript.model.templating.contextloader.ZscriptModuleTemplatingContextLoader</contextLoaderClass>
                     <fileTypeSuffix>hpp</fileTypeSuffix>
                 </configuration>
