--- util/mustache-templating/mustache-templating-test-resources/pom.xml
@@ -1,21 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>net.zscript.maven-templates</groupId>
-        <artifactId>mustache-templating</artifactId>
-        <version>0.0.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>mustache-templating-maven-plugin-test-resources</artifactId>
-    <name>Maven Templating Plugin Classpath Test Resources</name>
-
-    <description>
-        Testing only - don't make this a dependency of any other projects!
-        The goal of this project is to provide classpath resources that the
-        mustache-templating-maven-plugin-tests project can try to load.
-    </description>
-
-</project>
