--- util/mustache-templating/mustache-templating-test-resources/src/main/resources/template-cp/foo-4.mustache
@@ -1,1 +1,0 @@
-foo-4.mustache: - Ship-to: {{#ship-to}}{{street}}{{/ship-to}}
