--- util/mustache-templating/mustache-templating-test-resources/src/main/resources/template-cp/test-2-cp.mustache
@@ -1,1 +1,0 @@
-Test-2 (classpath): Test mustache file: receipt is {{receipt}} for {{#customer}}{{first_name}} {{family_name}}{{/customer}}
