--- util/mustache-templating/mustache-templating-test-resources/src/main/resources/template-cp/test-4-cp-part.mustache
@@ -1,3 +1,0 @@
-Test-4 (classpath, partials): Test mustache file: receipt is {{receipt}} for {{#customer}}{{first_name}} {{family_name}}{{/customer}}
-Testing partials:
-{{> foo-4.mustache}}
