--- util/mustache-templating/mustache-templating-tests/src/main/templates/test-1.mustache
@@ -1,1 +1,0 @@
-Test-1: Test mustache file: receipt is {{receipt}} for {{#customer}}{{first_name}} {{family_name}}{{/customer}}
