--- util/mustache-templating/mustache-templating-tests/src/main/templates/test3/partials/test-3a.mustache
@@ -1,1 +1,0 @@
-Test-3a: Bill-to: {{#bill-to}}{{street}}{{/bill-to}}
