--- util/mustache-templating/mustache-templating-tests/src/main/templates/test3/partials/test-3b.mustache
@@ -1,1 +1,0 @@
-Test-3b: - Ship-to: {{#ship-to}}{{street}}{{/ship-to}}
