--- util/mustache-templating/mustache-templating-tests/src/main/templates/test3/test-3.mustache
@@ -1,3 +1,0 @@
-Test-3: Test mustache file: receipt is {{receipt}} for {{#customer}}{{first_name}} {{family_name}}{{/customer}}
-Partial#1: {{> partials/test-3a.mustache}}
-Partial#2: {{> partials/test-3b.mustache}}
