--- util/mustache-templating/pom.xml
@@ -1,47 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>net.zscript</groupId>
-        <artifactId>zscript-util</artifactId>
-        <version>0.0.1-SNAPSHOT</version>
-    </parent>
-
-    <groupId>net.zscript.maven-templates</groupId>
-    <artifactId>mustache-templating</artifactId>
-    <packaging>pom</packaging>
-    <name>Mustache Templating Plugin Package</name>
-
-    <properties>
-        <version.mustache>0.9.10</version.mustache>
-    </properties>
-
-    <modules>
-        <module>mustache-templating-maven-plugin</module>
-        <module>mustache-templating-maven-plugin-adapter</module>
-        <module>mustache-templating-tests</module>
-        <module>mustache-templating-test-resources</module>
-    </modules>
-
-    <dependencyManagement>
-        <dependencies>
-            <dependency>
-                <groupId>com.github.spullara.mustache.java</groupId>
-                <artifactId>compiler</artifactId>
-                <version>${version.mustache}</version>
-            </dependency>
-        </dependencies>
-    </dependencyManagement>
-
-    <dependencies>
-        <dependency>
-            <groupId>net.zscript</groupId>
-            <artifactId>java-testing-deps</artifactId>
-            <version>${project.version}</version>
-            <scope>test</scope>
-            <type>pom</type>
-        </dependency>
-    </dependencies>
-</project>
