+++ util/pom.xml
@@ -15,7 +15,6 @@
 
     <modules>
         <module>misc</module>
-        <module>mustache-templating</module>
         <module>java-testing-deps</module>
     </modules>
 </project>
